users/requirements.txt
alabaster==0.7.10
alembic==0.9.7
aniso8601==2.0.0
attrs==17.4.0
Babel==2.5.3
bcrypt==3.1.4
behave==1.2.5
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.2.1
docutils==0.14
extras==1.0.0
fixtures==3.0.0
Flask==0.12.2
Flask-BasicAuth==0.2.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
flask-restful-swagger-2==0.35
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
fuzzywuzzy==0.16.0
gunicorn==19.7.1
idna==2.6
imagesize==1.0.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
linecache2==1.0.0
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
mysql-connector-python==8.0.6
nose==1.3.0
packaging==17.1
parse==1.8.2
parse-type==0.4.2
parso==0.1.1
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.4.3
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.2.0
pytest==3.4.0
pytest-cov==2.5.1
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
python-mimeparse==1.6.0
python-subunit==1.2.0
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.1
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
splinter==0.7.7
SQLAlchemy==1.2.2
sure==1.4.7
testtools==2.3.0
tox==2.9.1
traceback2==1.4.0
traitlets==4.3.2
unittest2==1.1.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.14.1